+++ packages/site/src/view/index/components/carousel.vue
